requirements.txt
streamlit
math
pyargus.directionEstimation
scipy.spatial
collections
numpy